Refactor integration tests into different modules
